Minor optimization of Resample, fixing cases where we were traversing more pixels than necessary, and while I'm at it, pulling the calls to floor/ceil out of the inner loop.

Performance results are:
```
- testPerfInseparableAwkwardSize (GafferImageTest.ResampleTest.ResampleTest) : was 2.00s now 0.74s (63% reduction)
- testPerfInseparableDisk (GafferImageTest.ResampleTest.ResampleTest) : was 6.31s now 5.92s (6% reduction)
- testPerfInseparableLanczos (GafferImageTest.ResampleTest.ResampleTest) : was 15.16s now 14.44s (5% reduction)
- testPerfSmallFilter (GafferImageTest.ResampleTest.ResampleTest) : was 0.43s now 0.39s (8% reduction)
- testPerfVerySmallFilter (GafferImageTest.ResampleTest.ResampleTest) : was 0.39s now 0.36s (10% reduction)
```

Also, now that a box filter just samples the values it uses, rather than also taking adjacent pixels and multiplying them by zero, it no longer converts `inf` to `NaN`, so we can remove the workaround from MergeTest.